move docid processing code to Presentation XML: https://github.com/metanorma/isodoc/issues/606
